Moved all the Table related code to the Table folder.
