+++ tools/gulptasks/dashboards/api-docs.json
@@ -71,11 +71,11 @@
         // DataGrid
         "../../../ts/DataGrid/DataGrid.ts",
         "../../../ts/DataGrid/Options.ts",
-        "../../../ts/DataGrid/Cell.ts",
-        "../../../ts/DataGrid/Column.ts",
-        "../../../ts/DataGrid/Row.ts",
-        "../../../ts/DataGrid/Table.ts",
-        "../../../ts/DataGrid/TableHead.ts",
+        "../../../ts/DataGrid/Table/Cell.ts",
+        "../../../ts/DataGrid/Table/Column.ts",
+        "../../../ts/DataGrid/Table/Row.ts",
+        "../../../ts/DataGrid/Table/Table.ts",
+        "../../../ts/DataGrid/Table/Header/TableHeader.ts",
         "../../../ts/DataGrid/Globals.ts",
         "../../../ts/DataGrid/Utils.ts",
 
@@ -88,10 +88,10 @@
         "../../../ts/DataGrid/Content/TableCell.ts",
         "../../../ts/DataGrid/Content/TableRow.ts",
 
-        "../../../ts/DataGrid/Actions/CellEditing.ts",
-        "../../../ts/DataGrid/Actions/ColumnsResizer.ts",
-        "../../../ts/DataGrid/Actions/ColumnSorting.ts",
-        "../../../ts/DataGrid/Actions/RowsVirtualizer.ts",
+        "../../../ts/DataGrid/Table/Actions/CellEditing.ts",
+        "../../../ts/DataGrid/Table/Actions/ColumnsResizer.ts",
+        "../../../ts/DataGrid/Table/Actions/ColumnSorting.ts",
+        "../../../ts/DataGrid/Table/Actions/RowsVirtualizer.ts",
     ],
     "entryPointStrategy": "expand",
     "excludeInternal": true,
