+++ ts/Dashboards/datagrid.d.ts
@@ -1,5 +1,5 @@
 export { default as DataGrid } from '../DataGrid/DataGrid.js';
-export { default as Column } from '../DataGrid/Column.js';
-export { default as TableRow } from '../DataGrid/Content/TableRow.js';
-export { default as TableCell } from '../DataGrid/Content/TableCell.js';
+export { default as Column } from '../DataGrid/Table/Column.js';
+export { default as TableRow } from '../DataGrid/Table/Content/TableRow.js';
+export { default as TableCell } from '../DataGrid/Table/Content/TableCell.js';
 export { default as Options } from '../DataGrid/Options.js';
