+++ ts/DataGrid/Table/Table.ts
@@ -22,18 +22,18 @@
  *
  * */
 
-import type { ColumnDistribution } from './Options';
+import type { ColumnDistribution } from '../Options';
 import type TableCell from './Content/TableCell';
 
-import DGUtils from './Utils.js';
-import DataTable from '../Data/DataTable.js';
+import DGUtils from '../Utils.js';
+import DataTable from '../../Data/DataTable.js';
 import Column from './Column.js';
 import TableHeader from './Header/TableHeader.js';
-import DataGrid from './DataGrid.js';
+import DataGrid from '../DataGrid.js';
 import RowsVirtualizer from './Actions/RowsVirtualizer.js';
 import ColumnsResizer from './Actions/ColumnsResizer.js';
-import Globals from './Globals.js';
-import Utils from '../Core/Utilities.js';
+import Globals from '../Globals.js';
+import Utils from '../../Core/Utilities.js';
 import CellEditing from './Actions/CellEditing.js';
 import TableRow from './Content/TableRow';
